+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+# [3.8.0](https://github.com/xability/maidr/compare/v3.7.1...v3.8.0) (2025-05-29)
+
+
+### Features
+
+* show settings button when no agents are enabled ([#248](https://github.com/xability/maidr/issues/248)) ([27ba2ca](https://github.com/xability/maidr/commit/27ba2cabf0bf616bc6dccb45dc83798769de0b7c))
+
+## [3.7.1](https://github.com/xability/maidr/compare/v3.7.0...v3.7.1) (2025-05-28)
+
+
+### Bug Fixes
+
+* optimize candlestick navigation with value-based sorting ([#286](https://github.com/xability/maidr/issues/286)) ([6dce5df](https://github.com/xability/maidr/commit/6dce5df281e0d117cfbdf29ee1cc293a9ccf49ec))
+
+# [3.7.0](https://github.com/xability/maidr/compare/v3.6.0...v3.7.0) (2025-05-28)
+
+
+### Features
+
+* enhance navigation with y-value-based movement for UPWARD/DOWNWARD directions ([#285](https://github.com/xability/maidr/issues/285)) ([4b94f17](https://github.com/xability/maidr/commit/4b94f176e31d3c0807663335169986091dfc5834))
+
+# [3.6.0](https://github.com/xability/maidr/compare/v3.5.0...v3.6.0) (2025-05-28)
+
+
+### Features
+
+* let users choose specific llm models within providers ([#253](https://github.com/xability/maidr/issues/253)) ([7803099](https://github.com/xability/maidr/commit/7803099358e96cce6622daf68700f5f2783c682b))
+
+# [3.5.0](https://github.com/xability/maidr/compare/v3.4.1...v3.5.0) (2025-05-27)
+
+
+### Features
+
+* support audio palette for multiclass plots ([#279](https://github.com/xability/maidr/issues/279)) ([3d70841](https://github.com/xability/maidr/commit/3d7084147130ac5c75097f279456e2f3186f2faa))
+
 ## [3.4.1](https://github.com/xability/maidr/compare/v3.4.0...v3.4.1) (2025-05-25)